Make storage the ServiceDirectory strategy: one concrete directory in core, repository SPI for modules
ServiceDirectory is the consumer API only — the liveness write methods
move to ServiceDirectoryRepository, the storage strategy a module
implements to get the full directory (capture, verification, liveness
maintenance) from DefaultServiceDirectory. kinotic-domain keeps just its
Elasticsearch repository and wires the bean; AbstractServiceDirectory and
ElasticServiceDirectory dissolve.
